backend/requirements.txt
Django==4.2
djangorestframework==3.20.4
django-cors-headers==3.15.0
